+++ source/lib/tests/test_prod_virial_a.cc
@@ -118,10 +118,10 @@ class TestProdVirialA : public ::testing::Test {
     }
     build_nlist(nlist_a_cpy, nlist_r_cpy, posi_cpy, nloc, rc, rc, nat_stt,
                 ncell, ext_stt, ext_end, region, ncell);
-    nlist.resize(nloc * nnei);
-    env.resize(nloc * ndescrpt);
-    env_deriv.resize(nloc * ndescrpt * 3);
-    rij.resize(nloc * nnei * 3);
+    nlist.resize(static_cast<size_t>(nloc) * nnei);
+    env.resize(static_cast<size_t>(nloc) * ndescrpt);
+    env_deriv.resize(static_cast<size_t>(nloc) * ndescrpt * 3);
+    rij.resize(static_cast<size_t>(nloc) * nnei * 3);
     for (int ii = 0; ii < nloc; ++ii) {
       // format nlist and record
       format_nlist_i_cpu<double>(fmt_nlist_a, posi_cpy, atype_cpy, ii,
@@ -144,7 +144,7 @@ class TestProdVirialA : public ::testing::Test {
         rij[ii * nnei * 3 + jj] = t_rij[jj];
       }
     }
-    net_deriv.resize(nloc * ndescrpt);
+    net_deriv.resize(static_cast<size_t>(nloc) * ndescrpt);
     for (int ii = 0; ii < nloc * ndescrpt; ++ii) {
       net_deriv[ii] = 10 - ii * 0.01;
     }
